ai_mimari/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy>=1.23.5
pillow>=10.0.0
werkzeug==2.0.1
torch==2.7.0
torchvision==0.22.0
